Adds an explanation page for job types
Adds an explanation of the different job types, and how they might be
used, and adds them to a new Topic Guides section.  This corresponds to
the fourth type of documentation (Explanation) and should contain
explanatory articles providing background and context for readers
interested in gaining understanding.

We probably want to also add a How-to page for using the different jobs
types.